move some of fieldpage to fieldpageskeleton
