+++ frontends/infinite-corridor/src/pages/index.ts
